+++ src/Components/Navbar/UserNavbar.js
@@ -1,9 +1,7 @@
-import React, { useState, useEffect } from 'react';
-import Cookies from 'universal-cookie';
+import React, { useEffect, useState } from 'react';
 import {
   ButtonDropdown,
   Collapse,
-  CustomInput,
   NavbarToggler,
   Navbar,
   NavbarBrand,
@@ -15,15 +13,12 @@ import {
   DropdownMenu,
   DropdownItem
 } from 'reactstrap';
+import DarkMode from './DarkMode';
 import { membershipState } from '../../Enums';
 import logo from '../Navbar/sce_logo.png';
-import { toggleDarkTheme } from '../../APIFunctions/dark-theme';
-import { sunIcon, moonIcon } from '../../Pages/Overview/SVG';
-
 
 export default function UserNavBar(props) {
   const [menuIsOpen, setMenuIsOpen] = useState(false);
-  const [darkTheme, setDarkTheme] = useState(true);
   const icons = [
     {
       link: ['https://www.linkedin.com/company', '/sjsusce/'].join(''),
@@ -86,26 +81,7 @@ export default function UserNavBar(props) {
     setMenuIsOpen(!menuIsOpen);
   };
 
-  function createCookie() {
-    if (document.body.className === 'light') {
-      setDarkTheme(true);
-    } else {
-      setDarkTheme(false);
-    }
-    toggleDarkTheme();
-  }
-
-  function keepState() {
-    const cookie = new Cookies();
-    if (cookie.get('dark') === 'true') {
-      setDarkTheme(true);
-    } else {
-      setDarkTheme(false);
-    }
-  }
-
   useEffect(() => {
-    keepState();
     // eslint-disable-next-line
   }, [])
 
@@ -117,15 +93,8 @@ export default function UserNavBar(props) {
             <img id='logo-image' src={logo} alt={'sce logo'} />
           </div>
         </NavbarBrand>
-        <div id='toggler-left'>
-          <div id='sun-icon' >{sunIcon()}</div>
-          <div id='dark-toggler'>
-            <CustomInput
-              onChange={createCookie} className='darkToggle'
-              type='switch' id='exampleCustomSwitch' name='customSwitch'
-              checked={darkTheme} />
-          </div>
-          <div id='moon-icon' >{moonIcon()}</div>
+        <div id='user-dark'>
+          <DarkMode />
         </div>
         <NavbarToggler tag='h1'>
           <ButtonDropdown isOpen={menuIsOpen} toggle={toggler}>
@@ -150,20 +119,20 @@ export default function UserNavBar(props) {
               </DropdownItem>
               {props.user && props.user.accessLevel >=
                 membershipState.MEMBER && (
-                  <DropdownItem tag='h1' className='dropdown-submenu drp-item'>
-                    <DropdownItem className='drp-item' id='btndrp-text'>
+                <DropdownItem tag='h1' className='dropdown-submenu drp-item'>
+                  <DropdownItem className='drp-item' id='btndrp-text'>
                       Printing
                   </DropdownItem>
-                    <DropdownMenu className='drp-menu'>
-                      <DropdownItem className='drp-item' href='/2DPrinting'>
+                  <DropdownMenu className='drp-menu'>
+                    <DropdownItem className='drp-item' href='/2DPrinting'>
                         2D Printing
                     </DropdownItem>
-                      <DropdownItem className='drp-item' href='/3DPrintingForm'>
+                    <DropdownItem className='drp-item' href='/3DPrintingForm'>
                         3D Printing
                     </DropdownItem>
-                    </DropdownMenu>
-                  </DropdownItem>
-                )}
+                  </DropdownMenu>
+                </DropdownItem>
+              )}
               {unauthedRoutes.map((link, index) => {
                 return (
                   <DropdownItem className='drp-item' key={index}>
@@ -216,20 +185,20 @@ export default function UserNavBar(props) {
                   </DropdownMenu>
                 </DropdownItem>
               ) : (
-                  <DropdownItem tag='h1' className='dropdown-submenu drp-item'>
-                    <DropdownItem className='drp-item' id='btndrp-text'>
+                <DropdownItem tag='h1' className='dropdown-submenu drp-item'>
+                  <DropdownItem className='drp-item' id='btndrp-text'>
                       Join Us!
                   </DropdownItem>
-                    <DropdownMenu right className='drp-menu'>
-                      <DropdownItem className='drp-item' href='/register'>
+                  <DropdownMenu right className='drp-menu'>
+                    <DropdownItem className='drp-item' href='/register'>
                         Membership Application
                     </DropdownItem>
-                      <DropdownItem className='drp-item' href='/login'>
+                    <DropdownItem className='drp-item' href='/login'>
                         Login
                     </DropdownItem>
-                    </DropdownMenu>
-                  </DropdownItem>
-                )}
+                  </DropdownMenu>
+                </DropdownItem>
+              )}
             </DropdownMenu>
           </ButtonDropdown>
         </NavbarToggler>
@@ -309,23 +278,23 @@ export default function UserNavBar(props) {
                 </DropdownMenu>
               </UncontrolledDropdown>
             ) : (
-                <UncontrolledDropdown nav inNavbar>
-                  <DropdownToggle id='navlink-text' nav caret>
+              <UncontrolledDropdown nav inNavbar>
+                <DropdownToggle id='navlink-text' nav caret>
                     Join Us!
                 </DropdownToggle>
-                  <DropdownMenu right>
-                    <DropdownItem className='drp-item' href='/register'>
+                <DropdownMenu right>
+                  <DropdownItem className='drp-item' href='/register'>
                       Membership Application
                   </DropdownItem>
-                    <DropdownItem className='drp-item' href='/login'>
+                  <DropdownItem className='drp-item' href='/login'>
                       Login
                   </DropdownItem>
-                  </DropdownMenu>
-                </UncontrolledDropdown>
-              )}
+                </DropdownMenu>
+              </UncontrolledDropdown>
+            )}
           </Nav>
         </Collapse>
       </Navbar>
-    </div>
+    </div >
   );
 }
